♻️ Refactor: 커스텀 훅의 형태로 container수정
