+++ src/containers/signup-pagea/useSignUpPage.tsx
@@ -7,7 +7,7 @@ import emailBlur from "@/src/utils/sign-blur-error-message/input-blur-types/emai
 import passwordBlur from "@/src/utils/sign-blur-error-message/input-blur-types/passwordBlur";
 import passCheckBlur from "@/src/utils/sign-blur-error-message/input-blur-types/passCheckBlur";
 
-export default function SignUpPageContainer() {
+export default function useSignUpPage() {
   const [emailError, setEmailError] = useState("");
   const [passwordError, setPasswordError] = useState("");
   const [passCheckError, setPassCheckError] = useState("");
